fix update

+++ src/CoreBundle/Bridge/Chameleon/Migration/Script/update-1758609692.inc.php
@@ -5,6 +5,15 @@
 </div>
 <?php
 
-$query = "ALTER TABLE cms_media DROP INDEX metatags";
+$indexExists = TCMSLogChange::getDatabaseConnection()->fetchOne("
+    SELECT COUNT(1)
+    FROM INFORMATION_SCHEMA.STATISTICS
+    WHERE TABLE_SCHEMA = DATABASE()
+    AND TABLE_NAME = ?
+    AND INDEX_NAME = ?
+", ['cms_media', 'metatags']);
 
-TCMSLogChange::RunQuery(__LINE__, $query);
+if ($indexExists) {
+    $query = "ALTER TABLE cms_media DROP INDEX metatags";
+    TCMSLogChange::RunQuery(__LINE__, $query);
+}
